more size_t

+++ test/profiler_manager_iterations_test.cc
@@ -38,7 +38,7 @@ int main(int argc, char** argv) {
   // Make a fake argv and append the new --benchmark_profiler_iterations=<foo>
   // to it.
   int fake_argc = argc + 1;
-  std::vector<const char*> fake_argv(fake_argc);
+  std::vector<const char*> fake_argv(static_cast<size_t>(fake_argc));
   for (size_t i = 0; i < static_cast<size_t>(argc); ++i) {
     fake_argv[i] = argv[i];
   }
